requirements_mcp.txt
# Travel Planning MCP Server Dependencies

# Core MCP and async dependencies
mcp>=1.0.0
asyncio-mqtt>=0.16.0
aiofiles>=23.2.1

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Google AI (Gemini) integration
google-generativeai>=0.3.2
google-cloud-aiplatform>=1.38.0

# CrewAI and related AI libraries
crewai>=0.41.1
crewai-tools>=0.4.26
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.1

# Web scraping and search tools
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.0

# Data processing and utilities
pandas>=2.1.0
numpy>=1.24.0
python-dateutil>=2.8.2
pytz>=2023.3

# Logging and monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Configuration and environment
python-dotenv>=1.0.0
PyYAML>=6.0.1
toml>=0.10.2

# Testing dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development dependencies (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Optional travel APIs (uncomment if using real APIs)
# amadeus>=8.0.0
# googlemaps>=4.10.0

# Cache and performance
redis>=5.0.0
diskcache>=5.6.3

# HTTP client improvements
httpx>=0.25.0
aiohttp>=3.9.0